Langchain Projects/Advanced_Rag/requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
faiss-cpu
langchain_text_splitters
wikipedia
arxiv
langchainhub
